rename browser to runner
